+++ src/ZPublisher/cookie.py
@@ -25,6 +25,8 @@ and exports the functions ``getCookieParamPolicy``, ``getCookieValuePolicy``
 ``normalizeCookieParameterName`` and ``convertCookieParameter``.
 """
 import datetime
+from encodings.idna import ToASCII
+from encodings.idna import nameprep
 from itertools import chain
 from re import compile
 from time import time
@@ -241,8 +243,11 @@ def domain_converter(value):
     u_value = value.decode("utf-8") if isinstance(value, bytes) else value
     if "xn--" in u_value:  # already encoded
         return value
-    from encodings.idna import ToASCII
-    from encodings.idna import nameprep
+
+    # According to https://www.rfc-editor.org/rfc/rfc6265#section-4.1.2.3 a
+    # leading dot is ignored. If it is there `ToASCII`, breaks on the empty
+    # string:
+    u_value = u_value.lstrip('.')
     return ".".join(to_str(ToASCII(nameprep(c))) for c in u_value.split("."))